# backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: EduPilot Gateway API
  version: 0.14.0
  description: |
    Hợp đồng API của gateway Go — NGUỒN SỰ THẬT (D45). Mỗi route trong mã phải có ở đây và ngược lại
    (`internal/contract` kiểm hai chiều). Route thử nằm ở `openapi.test.yaml`, chỉ có trong binary dựng bằng
    build tag `testroutes`. Phase PG chưa có endpoint nghiệp vụ; P1 (FEAT-llm-gateway US-P1-04) thêm 13 thao tác `/admin/llm/*` và `/courses/{id}/llm-budget`.
    P2 (FEAT-account-security US-P2-02) thêm 9 thao tác `/auth/{login,refresh,logout,register,verify-email,resend-verification,forgot-password,reset-password,tokens/preview}` và 4 thao tác `/me/{password,sessions,sessions/{id}}` (US-P2-04); US-P2-05 thêm 429 (`LOGIN_THROTTLED` | `RATE_LIMITED`) cho login / register / verify-email / reset-password / refresh: công khai (không Bearer); `refresh`/`logout` dùng cookie `ep_rt` (HttpOnly) + kiểm Origin.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/courses/{id}/questions:
    get:
      operationId: listQuestions
      tags: [exam]
      summary: Danh sách câu hỏi của lớp (Giảng viên / TA; lọc, phân trang con trỏ)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: review_status
          in: query
          schema:
            type: string
        - name: topic
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: origin
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: archived
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        "200":
          description: Trang câu hỏi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    post:
      operationId: createQuestion
      tags: [exam]
      summary: Tạo câu hỏi (DRAFT; MCQ_SINGLE, MCQ_MULTI, TRUE_FALSE, CODE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Câu đã tạo (kèm details.warnings nếu có)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/questions/suggest:
    post:
      operationId: suggestQuestions
      tags: [exam]
      summary: AI gợi ý nháp câu trắc nghiệm hoặc đầu vào test (việc nền, làn BATCH)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestRequest"
      responses:
        "202":
          description: Đã xếp việc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/questions/{qid}:
    get:
      operationId: getQuestion
      tags: [exam]
      summary: Chi tiết câu hỏi đầy đủ cho Staff (đáp án, giải thích, bộ test, lời giải mẫu)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Câu hỏi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateQuestion
      tags: [exam]
      summary: Sửa câu hỏi (khoá lạc quan theo version; câu đang dùng chỉ sửa chủ đề / độ khó / giải thích)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: Câu sau khi sửa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/questions/{qid}/archive:
    post:
      operationId: archiveQuestion
      tags: [exam]
      summary: Lưu trữ câu hỏi (không xoá cứng)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Câu đã lưu trữ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/courses/{id}/questions/{qid}/duplicate:
    post:
      operationId: duplicateQuestion
      tags: [exam]
      summary: Nhân bản câu hỏi thành bản DRAFT (kể cả test và lời giải mẫu)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Bản sao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/courses/{id}/questions/{qid}/review:
    put:
      operationId: reviewQuestion
      tags: [exam]
      summary: Gửi duyệt / duyệt / từ chối câu hỏi (REQUEST | APPROVE | REJECT)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
      responses:
        "200":
          description: Câu sau khi đổi trạng thái
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/questions/{qid}/code:
    put:
      operationId: putQuestionCode
      tags: [exam]
      summary: "Cấu hình bài lập trình: ngôn ngữ, giới hạn, checker, mã khởi tạo, lời giải mẫu"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeConfigRequest"
      responses:
        "200":
          description: Câu sau khi đổi cấu hình
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/questions/{qid}/testcases:
    get:
      operationId: listTestcases
      tags: [exam]
      summary: Danh sách test của bài (cả test ẩn, cả test AI chưa duyệt)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        "200":
          description: Trang test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestcasePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    post:
      operationId: createTestcase
      tags: [exam]
      summary: "Thêm một test (tạo tay: đã duyệt, tăng tests_version)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestcaseRequest"
      responses:
        "201":
          description: Test đã thêm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testcase"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/questions/{qid}/testcases/import:
    post:
      operationId: importTestcases
      tags: [exam]
      summary: Nhập test bằng zip (multipart, trường file; ?dry_run, ?mode=append|replace; all-or-nothing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: dry_run
          in: query
          schema:
            type: boolean
        - name: mode
          in: query
          schema:
            type: string
            enum: [append, replace]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Báo cáo thử (dry_run)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZipImportReport"
        "201":
          description: Đã nhập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZipImportReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/ValidationFailed"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/questions/{qid}/testcases/approve:
    post:
      operationId: approveTestcases
      tags: [exam]
      summary: Duyệt hàng loạt test AI gợi ý (tăng tests_version, gỡ cờ lời giải mẫu đã kiểm)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveTestsRequest"
      responses:
        "200":
          description: Số test đã duyệt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveTestsResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/questions/{qid}/testcases/{tid}:
    put:
      operationId: updateTestcase
      tags: [exam]
      summary: Sửa một test (trường vắng giữ nguyên; đổi position chèn và đánh số lại)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestcaseRequest"
      responses:
        "200":
          description: Test sau khi sửa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testcase"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    delete:
      operationId: deleteTestcase
      tags: [exam]
      summary: Xoá một test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Đã xoá
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/courses/{id}/questions/{qid}/reference/verify:
    post:
      operationId: verifyReference
      tags: [exam]
      summary: Chạy lời giải mẫu qua mọi test đã duyệt (việc nền; kết quả ở jobs.result)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: qid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Đã xếp việc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Chưa đăng nhập, token hết hạn / không hợp lệ / phiên bị thu hồi (`UNAUTHENTICATED`, `TOKEN_EXPIRED`, `TOKEN_INVALID`, `SESSION_REVOKED`, `INVALID_CREDENTIALS`)
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Sai vai trò, ngoài lớp hoặc Origin lạ (`FORBIDDEN`, `details.reason` = `role` | `course` | `origin`); tài khoản bị khoá (`ACCOUNT_DISABLED`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Không có route / tài nguyên / không có quyền biết (`NOT_FOUND`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationFailed:
      description: Dữ liệu không hợp lệ (`VALIDATION_FAILED`, `INVALID_CURSOR`, `IDEMPOTENCY_KEY_*`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    QuestionInput:
      type: object
      required: [type, title, topic, stem]
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [MCQ_SINGLE, MCQ_MULTI, TRUE_FALSE, CODE, SHORT, ESSAY]
        title:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        stem:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/OptionInput"
        correct:
          type: array
          items:
            type: integer
        value:
          type: [boolean, "null"]
        explanation:
          type: [string, "null"]
        version:
          type: integer
...
    QuestionPage:
      type: object
      required: [items, next_cursor]
      additionalProperties: false
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/QuestionListItem"
        next_cursor:
          type: [string, "null"]
...